refa: initializeStore에 직접 createAppStore 로직 추가
